requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
boto3==1.34.0
pydantic==2.6.0
python-multipart==0.0.6
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
